project done

+++ src/pages/article/Employeearticle.js
@@ -1,6 +1,6 @@
 import React from "react";
 import { useState, useEffect } from "react";
-import { NavLink, useHistory, useRouteMatch, Route,Redirect } from "react-router-dom";
+import { NavLink, useHistory, useRouteMatch, Route, Redirect } from "react-router-dom";
 import { styled } from "@mui/material/styles";
 import {
   Typography,
@@ -49,28 +49,31 @@ const StyledTableRow = styled(TableRow)(({ theme }) => ({
 }));
 const Employeearticle = () => {
   const navigate = useHistory();
-  const [userdata,setuserdata] = useState([]);
+  const [userdata, setuserdata] = useState([]);
   let { path, url } = useRouteMatch();
   const [articledata, setarticledata] = useState([]);
   const [counts, setcounts] = useState([]);
   const [page, setPage] = React.useState(0);
   const [rowsPerPage, setRowsPerPage] = React.useState(5);
 
-  const getLocalItem = async() => {
-    let udata=localStorage.getItem("user");
+  const getLocalItem = async () => {
+    let udata = await JSON.parse(localStorage.getItem("user"));
     if (udata) {
-      setuserdata(JSON.parse(udata));
+      setuserdata(udata);
     }
-    let data = await JSON.parse(localStorage.getItem("article"));
+  };
+  const getArticleData = async () => {
+    let data =JSON.parse(localStorage.getItem("article"));
     if (data) {
-      let adata=data.filter(ad=>ad.isDeleted==false && ad.empid==userdata[0].id)
-      setarticledata( adata);
+      let adata =data.filter(ad => ad.isDeleted == false)
+      setarticledata(adata);
     } else {
       return [];
     }
-    console.log(articledata)
-  };
-
+  }
+  useEffect(()=>{
+    getArticleData();
+  },[userdata])
   useEffect(() => {
     try {
       getLocalItem();
@@ -91,45 +94,44 @@ const Employeearticle = () => {
     var date = new Date(str),
       mnth = ("0" + (date.getMonth() + 1)).slice(-2),
       day = ("0" + date.getDate()).slice(-2);
-    return [day,mnth,date.getFullYear()].join("-");
+    return [day, mnth, date.getFullYear()].join("-");
   }
   const handleDelete = (e) => {
     console.log(e);
-    const index = articledata.indexOf(articledata.find((a => a.id == e )));
+    const index = articledata.indexOf(articledata.find((a => a.id == e)));
     console.log(index)
     articledata[index].isDeleted = true
     setarticledata(articledata);
     console.log(articledata)
     navigate.push("/employeedashboard/employeearticle")
-    localStorage.setItem("article",JSON.stringify(articledata))
+    localStorage.setItem("article", JSON.stringify(articledata))
   }
-  console.log(articledata);
   // const countitems =announcementdata.filter(e => e.isDeleted ==='false').length;
   // console.log("Not deleted: " + countitems);
   return (
     <div className="page-information-container">
       <div className="page-header">
-        <label>Sales/Purchase</label>
+        <label>Article</label>
       </div>
       <div className='page-breadscrumb'>
-            <br/>
-              <Breadcrumbs aria-label="breadcrumb">
-                <Link underline="hover" color="inherit" href="/employeedashboard" exact to="/employeedashboard/">
-                  Home
-                </Link>
-                <Link
-                  underline="hover" color="inherit" href="/employeedashboard/employeearticle" exact to="/employeedashboard/employeearticle">
-                  Sale/Purchase
-                </Link>
-              </Breadcrumbs>
-              <br/>
-            </div>
+        <br />
+        <Breadcrumbs aria-label="breadcrumb">
+          <Link underline="hover" color="inherit" href="/employeedashboard" exact to="/employeedashboard/">
+            Home
+          </Link>
+          <Link
+            underline="hover" color="inherit" href="/admindashboard/employeearticle" exact to="/admindashboard/employeearticle">
+            Article
+          </Link>
+        </Breadcrumbs>
+        <br />
+      </div>
       <div className="upcomingevent-container">
         <div className="upcomingevent-container-button">
           {/* <Route exact path="/employeedashboard/employeesalespurchase/createemployeesalespurchase">
             <Createsalespurchase />
           </Route> */}
-          <NavLink to="/employeedashboard/employeearticle/createemployeearticle">
+          <NavLink to="/admindashboard/article/createarticle">
             <Button variant="contained" color="success" size="small">
               Create Article
             </Button>
@@ -155,7 +157,7 @@ const Employeearticle = () => {
                     <Checkbox
                       size="small"
                       name="upcomingeventSelect"
-                      sx={{ color:"white" }}></Checkbox>
+                      sx={{ color: "white" }}></Checkbox>
                   </StyledTableCell>
                   <StyledTableCell>Article Title</StyledTableCell>
                   <StyledTableCell>Article Link</StyledTableCell>
@@ -166,42 +168,43 @@ const Employeearticle = () => {
                   <StyledTableCell>Actions</StyledTableCell>
                 </TableRow>
               </TableHead>
-              { articledata.length > 0 &&
-              <TableBody>
-                {articledata
-                  .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
-                  .map((e) => {if(e.isDeleted == false)
-                    return (
-                      <StyledTableRow
-                        hover
-                        role="checkbox"
-                        tabIndex={-1}
-                        key={e.id}>
-                        <StyledTableCell>
-                          <Checkbox size="small" />
-                        </StyledTableCell>
-                        <StyledTableCell>{e.articleTitle}</StyledTableCell>
-                        <StyledTableCell>{e.articleLink}</StyledTableCell>
-                        <StyledTableCell>{e.articleDescription}</StyledTableCell>
-                        <StyledTableCell>{e.articleStartDate}</StyledTableCell>
-                        <StyledTableCell><img src={e.articleEndDate} height="50rem" width="50rem" style={{}} /></StyledTableCell>
-                        <StyledTableCell>{e.isApproved==false?"Not Approved":"Approved"}</StyledTableCell>
-                        <StyledTableCell>
-                          <LinkRoute
-                            to={{
-                              pathname:
-                                "/employeedashboard/employeearticle/updateemployeearticle/:id",
-                              state: { idParam: e.id },
-                            }}>
-                            <ModeEdit sx={{ color: "rgba(0, 127, 255, 1)" }} />
-                          </LinkRoute>
+              {articledata.length > 0 &&
+                <TableBody>
+                  {articledata
+                    .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
+                    .map((e) => {
+                      if (e.isDeleted == false)
+                        return (
+                          <StyledTableRow
+                            hover
+                            role="checkbox"
+                            tabIndex={-1}
+                            key={e.id}>
+                            <StyledTableCell>
+                              <Checkbox size="small" />
+                            </StyledTableCell>
+                            <StyledTableCell>{e.articleTitle}</StyledTableCell>
+                            <StyledTableCell><a href={e.articleLink} target="_blank">{e.articleLink}</a></StyledTableCell>
+                            <StyledTableCell>{e.articleDescription}</StyledTableCell>
+                            <StyledTableCell>{e.articleStartDate}</StyledTableCell>
+                            <StyledTableCell>{e.articleEndDate}</StyledTableCell>
+                            <StyledTableCell>{e.isApproved == false ? "Not Approved" : "Approved"}</StyledTableCell>
+                            <StyledTableCell>
+                              <LinkRoute
+                                to={{
+                                  pathname:
+                                    "/employeedashboard/employeearticle/updateemployeearticle/:id",
+                                  state: { idParam: e.id },
+                                }}>
+                                <ModeEdit sx={{ color: "rgba(0, 127, 255, 1)" }} />
+                              </LinkRoute>
 
-                          <Button size='small' id={e.id} key={e.id} onClick={(event) => handleDelete(e.id)} sx={{ verticalAlign: "bottom", minWidth: "auto" }}><Delete sx={{ color: "red" }} /></Button>
-                        </StyledTableCell>
-                      </StyledTableRow>
-                    );
-                  })}
-              </TableBody>}
+                              <Button size='small' id={e.id} key={e.id} onClick={(event) => handleDelete(e.id)} sx={{ verticalAlign: "bottom", minWidth: "auto" }}><Delete sx={{ color: "red" }} /></Button>
+                            </StyledTableCell>
+                          </StyledTableRow>
+                        );
+                    })}
+                </TableBody>}
             </Table>
           </TableContainer>
         </div>
